python/requirements.txt
python-dotenv
langchain==0.3.25
langchain-core==0.3.75
langchain-openai==0.3.24
langgraph==0.6.6
pydantic==2.10.6
typer
